Changes following PR comments + Ensure that there is no name clash
