+++ Sources/Impl/PinLayoutImpl+Warning.swift
@@ -20,7 +20,12 @@
 #if os(iOS) || os(tvOS)
 import UIKit
 
-fileprivate var numberFormatter: NumberFormatter?
+fileprivate var numberFormatter: NumberFormatter = {
+    let numberFormatter = NumberFormatter()
+    numberFormatter.numberStyle = .decimal
+    numberFormatter.decimalSeparator = "."
+    return numberFormatter
+}()
 
 extension PinLayoutImpl {
     internal func pointContext(method: String, point: CGPoint) -> String {
@@ -109,10 +114,10 @@ extension PinLayoutImpl {
     internal func displayWarning(_ text: String) {
         var displayText = "\n👉 \(text)"
 
-        let x = numberFormattter().string(from: NSNumber(value: Float(view.frame.origin.x)))!
-        let y = numberFormattter().string(from: NSNumber(value: Float(view.frame.origin.y)))!
-        let width = numberFormattter().string(from: NSNumber(value: Float(view.frame.size.width)))!
-        let height = numberFormattter().string(from: NSNumber(value: Float(view.frame.size.height)))!
+        let x = numberFormatter.string(from: NSNumber(value: Float(view.frame.origin.x)))!
+        let y = numberFormatter.string(from: NSNumber(value: Float(view.frame.origin.y)))!
+        let width = numberFormatter.string(from: NSNumber(value: Float(view.frame.size.width)))!
+        let height = numberFormatter.string(from: NSNumber(value: Float(view.frame.size.height)))!
 
         displayText += "\n(Layouted view info: Type: \(viewName(view)), Frame: x: \(x), y: \(y), width: \(width), height: \(height))"
         
@@ -142,16 +147,6 @@ extension PinLayoutImpl {
 
     internal func insetsDescription(_ insets: UIEdgeInsets) -> String {
         return "UIEdgeInsets(top: \(insets.top), left: \(insets.left), bottom: \(insets.bottom), right: \(insets.right))"
-    }
-
-    fileprivate func numberFormattter() -> NumberFormatter {
-        if numberFormatter == nil {
-            numberFormatter = NumberFormatter()
-            numberFormatter!.numberStyle = .decimal
-            numberFormatter!.decimalSeparator = "."
-        }
-
-        return numberFormatter!
     }
 }
