+++ Sources/Impl/PinLayoutImpl.swift
@@ -83,7 +83,7 @@ class PinLayoutImpl: PinLayout {
         if #available(iOS 11.0, tvOS 11.0, *) {
             return view.safeAreaInsets
         } else {
-            return view.computeSafeAreaInsets()
+            return view.pinlayoutComputeSafeAreaInsets()
         }
     }
