+++ docs/objective_c.md
@@ -9,7 +9,7 @@ PinLayout can also be used from Objective-C. The PinLayout interface is slightly
 ###### Example 1:
 This example implement the PinLayout's Intro example using objective-c 
 
-<a href="https://github.com/mirego/PinLayout/blob/master/Example/PinLayoutSample/UI/Examples/Intro/IntroView.swift"><img src="pinlayout-intro-example.png"/></a>
+<a href="https://github.com/mirego/PinLayout/blob/master/Example/PinLayoutSample/UI/Examples/Intro/IntroView.swift"><img src="images/pinlayout_intro_example_iphonex.png"/></a>
 
 ```
 - (void) layoutSubviews {
